requirements.txt
altgraph==0.17.2
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
Babel==2.10.3
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
configparser==5.3.0
crayons==0.4.0
Django==4.1.5
djangorestframework==3.14.0
exceptiongroup==1.1.0
Flask==2.1.3
future==0.18.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.22.4
outcome==1.2.0
packaging==22.0
pandas==1.4.3
pefile==2022.5.30
Pillow==9.3.0
psycopg2==2.9.5
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.0
pytweening==1.0.4
pytz==2022.1
pywhatkit==5.4
pywin32-ctypes==0.2.0
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
selenium==3.141.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
threadpoolctl==3.1.0
tkcalendar==1.6.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.13
webdriver-manager==3.4.2
Werkzeug==2.1.2
wikipedia==1.4.0
wsproto==1.2.0
